Revert rename SockJSHandler to SockJS
