+++ vertx-web/src/main/java/io/vertx/ext/web/handler/sockjs/impl/BaseTransport.java
@@ -49,7 +49,7 @@ import io.vertx.ext.auth.VertxContextPRNG;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.handler.PlatformHandler;
 import io.vertx.ext.web.handler.SecurityPolicyHandler;
-import io.vertx.ext.web.handler.sockjs.SockJSOptions;
+import io.vertx.ext.web.handler.sockjs.SockJSHandlerOptions;
 import io.vertx.ext.web.handler.sockjs.SockJSSocket;
 import io.vertx.ext.web.handler.sockjs.Transport;
 import io.vertx.ext.web.impl.RoutingContextInternal;
@@ -70,17 +70,17 @@ class BaseTransport {
 
   protected final Vertx vertx;
   protected final LocalMap<String, SockJSSession> sessions;
-  protected final SockJSOptions options;
+  protected final SockJSHandlerOptions options;
 
   static final String COMMON_PATH_ELEMENT_RE = "\\/[^\\/\\.]+\\/([^\\/\\.]+)\\/";
 
-  public BaseTransport(Vertx vertx, LocalMap<String, SockJSSession> sessions, SockJSOptions options) {
+  public BaseTransport(Vertx vertx, LocalMap<String, SockJSSession> sessions, SockJSHandlerOptions options) {
     this.vertx = vertx;
     this.sessions = sessions;
     this.options = options;
   }
 
-  protected SockJSSession getSession(RoutingContext rc, SockJSOptions options, String sessionID, Handler<SockJSSocket> sockHandler) {
+  protected SockJSSession getSession(RoutingContext rc, SockJSHandlerOptions options, String sessionID, Handler<SockJSSocket> sockHandler) {
     return sessions.computeIfAbsent(sessionID, s -> new SockJSSession(vertx, sessions, rc, s, options, sockHandler));
   }
 
@@ -130,7 +130,7 @@ class BaseTransport {
     }
   }
 
-  static void setJSESSIONID(SockJSOptions options, RoutingContext rc) {
+  static void setJSESSIONID(SockJSHandlerOptions options, RoutingContext rc) {
     String cookies = rc.request().getHeader(COOKIE);
     if (options.isInsertJSESSIONID()) {
       //Preserve existing JSESSIONID, if any
@@ -175,7 +175,7 @@ class BaseTransport {
     }
   }
 
-  static PlatformHandler createInfoHandler(final SockJSOptions options, final VertxContextPRNG prng) {
+  static PlatformHandler createInfoHandler(final SockJSHandlerOptions options, final VertxContextPRNG prng) {
     final long offset = 2L << 30;
 
     return new PlatformHandler() {
@@ -205,7 +205,7 @@ class BaseTransport {
     rc.response().putHeader(CACHE_CONTROL, "no-store, no-cache, no-transform, must-revalidate, max-age=0");
   }
 
-  static SecurityPolicyHandler createCORSOptionsHandler(SockJSOptions options, String methods) {
+  static SecurityPolicyHandler createCORSOptionsHandler(SockJSHandlerOptions options, String methods) {
     return rc -> {
       if (LOG.isTraceEnabled()) {
         LOG.trace("In CORS options handler");
