+++ vertx-web/src/main/java/io/vertx/ext/web/handler/sockjs/impl/HtmlFileTransport.java
@@ -44,7 +44,7 @@ import io.vertx.core.shareddata.LocalMap;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.handler.PlatformHandler;
-import io.vertx.ext.web.handler.sockjs.SockJSOptions;
+import io.vertx.ext.web.handler.sockjs.SockJSHandlerOptions;
 import io.vertx.ext.web.handler.sockjs.SockJSSocket;
 
 import java.util.regex.Pattern;
@@ -90,7 +90,7 @@ class HtmlFileTransport extends BaseTransport {
 
   private final Handler<SockJSSocket> sockHandler;
 
-  HtmlFileTransport(Vertx vertx, Router router, LocalMap<String, SockJSSession> sessions, SockJSOptions options, Handler<SockJSSocket> sockHandler) {
+  HtmlFileTransport(Vertx vertx, Router router, LocalMap<String, SockJSSession> sessions, SockJSHandlerOptions options, Handler<SockJSSocket> sockHandler) {
     super(vertx, sessions, options);
 
     this.sockHandler = sockHandler;
