+++ vertx-web/src/main/java/io/vertx/ext/web/handler/sockjs/impl/JsonPTransport.java
@@ -46,7 +46,7 @@ import io.vertx.ext.web.RequestBody;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.handler.PlatformHandler;
-import io.vertx.ext.web.handler.sockjs.SockJSOptions;
+import io.vertx.ext.web.handler.sockjs.SockJSHandlerOptions;
 import io.vertx.ext.web.handler.sockjs.SockJSSocket;
 
 import java.util.regex.Pattern;
@@ -63,7 +63,7 @@ class JsonPTransport extends BaseTransport {
 
   private final Handler<SockJSSocket> sockHandler;
 
-  JsonPTransport(Vertx vertx, Router router, LocalMap<String, SockJSSession> sessions, SockJSOptions options, Handler<SockJSSocket> sockHandler) {
+  JsonPTransport(Vertx vertx, Router router, LocalMap<String, SockJSSession> sessions, SockJSHandlerOptions options, Handler<SockJSSocket> sockHandler) {
     super(vertx, sessions, options);
 
     this.sockHandler = sockHandler;
