+++ vertx-web/src/main/java/io/vertx/ext/web/handler/sockjs/impl/XhrTransport.java
@@ -49,7 +49,7 @@ import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.handler.PlatformHandler;
 import io.vertx.ext.web.handler.SecurityPolicyHandler;
-import io.vertx.ext.web.handler.sockjs.SockJSOptions;
+import io.vertx.ext.web.handler.sockjs.SockJSHandlerOptions;
 import io.vertx.ext.web.handler.sockjs.SockJSSocket;
 
 import java.util.Arrays;
@@ -75,7 +75,7 @@ class XhrTransport extends BaseTransport {
 
   private final Handler<SockJSSocket> sockHandler;
 
-  XhrTransport(Vertx vertx, Router router, LocalMap<String, SockJSSession> sessions, SockJSOptions options, Handler<SockJSSocket> sockHandler) {
+  XhrTransport(Vertx vertx, Router router, LocalMap<String, SockJSSession> sessions, SockJSHandlerOptions options, Handler<SockJSSocket> sockHandler) {
 
     super(vertx, sessions, options);
