+++ vertx-web/src/test/java/io/vertx/ext/web/handler/sockjs/SockJSErrorTest.java
@@ -173,7 +173,7 @@ public class SockJSErrorTest extends VertxTestBase {
       .addInboundPermitted(permittedOptions)
       .addOutboundPermitted(permittedOptions)
       .setPingTimeout(60000);  // should not interfere with test
-    return SockJS.create(vertx).bridge(bridgeOptions, new TestBridgeEventHandler());
+    return SockJSHandler.create(vertx).bridge(bridgeOptions, new TestBridgeEventHandler());
   }
