+++ vertx-web/src/test/java/io/vertx/ext/web/handler/sockjs/SockJSTestBase.java
@@ -49,10 +49,10 @@ abstract class SockJSTestBase extends VertxTestBase {
   }
 
   void startServers() throws Exception {
-    startServers(new SockJSOptions());
+    startServers(new SockJSHandlerOptions());
   }
 
-  void startServers(SockJSOptions options) throws Exception {
+  void startServers(SockJSHandlerOptions options) throws Exception {
     CountDownLatch latch = new CountDownLatch(1);
     vertx.deployVerticle(() -> new AbstractVerticle() {
       @Override
@@ -70,7 +70,7 @@ abstract class SockJSTestBase extends VertxTestBase {
 
         options.setHeartbeatInterval(2000);
         options.setRegisterWriteHandler(true);
-        SockJS sockJSHandler = SockJS.create(vertx, options);
+        SockJSHandler sockJSHandler = SockJSHandler.create(vertx, options);
         router.route("/test/*").subRouter(sockJSHandler.socketHandler(socketHandler.get()));
 
         vertx.createHttpServer(new HttpServerOptions().setPort(8080).setHost("localhost"))
